chore: general cleanups
* Added a missing null check.
* Added missing doc strings.
* Downgrade inode and dev numbers to 32 bits.
* Added some logging statements.

+++ fact-ebpf/src/bpf/types.h
@@ -33,8 +33,8 @@ typedef struct process_t {
 } process_t;
 
 typedef struct inode_key_t {
-  unsigned long long inode;
-  unsigned long long dev;
+  unsigned long inode;
+  unsigned long dev;
 } inode_key_t;
 
 typedef char inode_value_t;
